Add serialize tests and fix a nasty double free bug when adding new root nodes
Previously if we added an existing node to a new document we would set
the root of the document to the current address. Then we would
potentially move the node, but the root was still the old address even
if we copied the node.

So the node would be free'd twice and seg fault.

Now we always copy the node first if there is no root in the document
and make the root the copied node, so there is no double free.