Fix silly copy paste error
When converting doc_parse_raw I did not edit the variable names
properly, so the url was always set to be the same as the encoding.
Additionally the NA object returned when there was no URL was not
the right type, it was a CHARSXP instead of a length 1 STRSXP. Both
these issues are now (again) fixed.

Fixes #300